sbool() takes arg of type unit64_t, not int
was returning incorrect answer if only a high order bit was set

+++ src/util/string_util.h
@@ -27,7 +27,7 @@ extern "C" {
 // General
 //
 
-static inline const char * sbool(int val) {  return (val)  ? "true" : "false"; }
+static inline const char * sbool(uint64_t val) {  return (val)  ? "true" : "false"; }
 
 // A macro alternative to sbool()
 #define SBOOL(val) ( (val) ? "true" : "false" )
